+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Last Changes
+- [#8](https://github.com/LaxarJS/laxar_uikit/issues/8): The axInput control now respects the scope locale for on-focus formatting.
 - [#7](https://github.com/LaxarJS/laxar_uikit/issues/7): Fixed grid depends on position of class col-lg in attribute.
 - [#6](https://github.com/LaxarJS/laxar_uikit/issues/6): Now the margin-top for a button variant (sm, xs) will calculate depending on bootstrap variants.
 - [#5](https://github.com/LaxarJS/laxar_uikit/issues/5): Removed console.log statements from text-ellipsis control.